+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 - Enh #180: Option to add wiki pages to Space menu
 - Enh #62: Add print version
 - Enh #138: Add "Save" action to sidebar navigation
+- Enh #9: Add activities for edited Wiki pages
 
 
 1.5.3  (April 16, 2021)